fix: INPUT overlapping prompt
When using colored prompt, readline miscalulates the size of prompt.
It behaves that if you write more than the size of the terminal, the new letter will be displayed ON the prompt.
If you put a '\n' at the end of the prompt it resolves that.
You can also DELETE the color, but it's less pretty.